Add 'inference' module containing RealTimeInference and Benchmark
